tada

+++ src/news/comments/comments.controller.ts
@@ -1,4 +1,6 @@
-import { Body, Controller, Get, Param, Post, Redirect, Render } from '@nestjs/common';
+import { Body, Controller, Get, Param, Post, Redirect, Render, UseGuards, Request, UnauthorizedException } from '@nestjs/common';
+import { AuthGuard } from '@nestjs/passport';
+import { UsersEntity } from 'src/users/users.entity';
 import { CommentsEntity } from './comments.entity';
 import { CommentsService } from './comments.service';
 
@@ -6,31 +8,49 @@ import { CommentsService } from './comments.service';
 export class CommentsController {
   constructor(private commentsService: CommentsService) { }
 
+  @UseGuards(AuthGuard('jwt'))
   @Post('/delete/:id')
   @Redirect()
-  async delete(@Param('id') id: string) {
+  async delete(@Param('id') id: string, @Request() req) {
     const commentIdInt = parseInt(id);
-    const newsId = (await this.commentsService.findById(commentIdInt)).news.id;
+    const news = (await this.commentsService.findById(commentIdInt)).news;
+
+    if (news.user.id != req.user.id) throw new UnauthorizedException();
+
     await this.commentsService.delete(commentIdInt);
-    return { url: `/news/${newsId}` };
+    return { url: `/news/${news.id}` };
   }
 
+  @UseGuards(AuthGuard('jwt'))
   @Get('/edit/:id')
   @Render('edit-comment')
   async edit(
     @Param('id') id: string,
-  ): Promise<{ title: string; comment: CommentsEntity }> {
+    @Request() req,
+  ): Promise<{ title: string; comment: CommentsEntity; user: UsersEntity }> {
     const commentIdInt = parseInt(id);
     const comment = await this.commentsService.findById(commentIdInt);
-    return { title: 'Редактирование комментария', comment: comment };
+
+    if (comment.user.id != req.user.id) throw new UnauthorizedException();
+
+    return {
+      title: 'Редактирование комментария',
+      comment: comment,
+      user: req.user,
+    };
   }
 
+  @UseGuards(AuthGuard('jwt'))
   @Post('/edit/:id')
   @Redirect()
-  async modify(@Param('id') id: string, @Body() body) {
+  async modify(@Param('id') id: string, @Body() body, @Request() req) {
     const commentIdInt = parseInt(id);
+    const news = (await this.commentsService.findById(commentIdInt)).news;
+    console.log(news.user);
+    console.log(req.user);
+    if (news.user.id != req.user.id) throw new UnauthorizedException();
+
     await this.commentsService.update(commentIdInt, body.message);
-    const newsId = (await this.commentsService.findById(commentIdInt)).news.id;
-    return { url: `/news/${newsId}` };
+    return { url: `/news/${news.id}` };
   }
 }
